NO-TICKET: adding little observability to blood donations
