--- .devcontainer/devcontainer.json
@@ -1,13 +1,0 @@
-{
-  "name": "BloodConnect",
-  "image": "dev-image",
-  "customizations": {
-      "vscode": {
-          "extensions": [
-              "ms-python.python",
-              "dbaeumer.vscode-eslint",
-              "hashicorp.terraform"
-          ]
-      }
-  }
-}
