+++ iac/terraform/aws/eventbridge/variables.tf
@@ -22,3 +22,15 @@ variable "donation_status_manager_queue_arn" {
   type        = string
   description = "ARN of the donation status manager SQS queue"
 }
+
+variable "max_geohash_length" {
+  type        = number
+  description = "geohash length for calculation"
+  default     = 8
+}
+
+variable "max_geohash_storage" {
+  type        = number
+  description = "total geohash to store per file"
+  default     = 582542
+}
